Add input to query tagging progress by organisations
Previously we were producing a list of organisations from users that exist in
content-tagger. However, not every organisation that we want to surface tagging
progress will exist in this list, as some users will be tagging content that
belongs to multiple organisations.

Instead, we now provide a way to query the tagging progress by a list of
organisation slugs. This also provides a URL that can be shared around and
could represent a selection of organisations (e.g. transport, or education)
